Fix alarm when receive TRACE
Since nova will save and reraise a exception when processing rpc request.
sentry may receive two notification in this case, the first one contains
not exception payload, just the title of the exception, and the second one
is the real exception. What we want to alarm is the second one. This patch
fix this.

Change-Id: I8a7d0d4aeb49b4e70522048487798894da487eb4